Use closeSession on QredoClient during tests
* All QredoClient instances created directly during tests now have closeSession
called at end of test (if a instance variable) or at end of test method (if
local variable) to correctly clean up the client after use.  Actual
implementation of closeSession still to be completed, but will avoid the leak
of memory/resources that MQTT currently causes due to not terminating the
transport threads.