Fix race condition during dedupe
* Rendezvous Response dedupe was being performed in 3 steps:
     1.) Check response is not present in dedupe store.
     2.) Create conversation from response.
     3.) Add response to dedupe store.
Steps 1 and 3 were synchronised against the dedupe store.  During testing had a
1-off failure where a duplicate response wasn't being deduped correctly.
Appeared to be a race condition between steps 1 and 3 where a Rendezvous
response come in during step 2.  As a result the dedupe store was empty, and
let the response through.
* Have now merged steps 1 and 3 into step 1 so both the check/add are performed
under the single synchronisation lock, eliminating the potential for a race
condition.  Extending the lock over all 3 steps may have added potential for
deadlock/delays as step 2 does vault comms so is not necessarily a quick
process. Merging steps 1 and 3 is best.