Add (unused) signingCallback arg for external keys
- Authenticated Rendezvous needs to support externally generated keys (e.g.
X.509 certificates, RSA keys), and allow the same keypair to be used when using
prefixes (so single public key can authenticate mutiple rendezvous). Either
must be able to pass in or specify the public and private keys (which could be
certificate, or raw key), or delegate the signing to the SDK caller - allowing
the SDK to not be involved with key storage or trying to pass in dissimilar key
formats (Ed25519, X.509 certificates and PEM encoded RSA keypairs) into a
generic function.  This latter option is what the Java SDK does via a signing
callback.
- Defined a common block type (signDataBlock) and added to the necessary call
chain so that the Rendezvous Helper can delegate the signing to the block
- Made the necessary changes through the call stack and tests, but no use of
the new argument is made yet.