Add clean .gcda files step to Podfile
- When code profiling is enabled, sometimes after executing tests etc the
profiler has problems merging the results into the existing .gcda files
(possibly because new methods added, or code removed?).
- Solution found online is to clean out the .gcda files on build.  This was
added as a Shell Script Build Phase to all targets in the the QredoSDK_pods
project which covers the SDK and resolved the problem.
- However, for the LF/Crypto projects which uses Cocoapods, the project file
can be overwritten frequently when installing/updating the Pod, as apparently
needed when adding new files to these projects. Thus the need to repeatedly add
this Shell Script step.
- This commit adds this Shell Script step to the Podfile, thus automatically
adding it whenever installing/updating the Pod