Move TrustedRootRefs out of CryptoImpl
- TrustedRootRefs was a member of CryptoImpl, set via the initialiser.
However, in a number of places the default initialiser for CryptoImpl was used,
including via a singleton 'instance' member within QredoRendezvousCrypto. This
meant that in most places it was used, TrustedRootRefs was always empty, thus
meaning X.509 authenticated rendezvous would always fail validation.
- TrustedRootRefs is now passed in when creating, or responding to a
rendezvous. This allows the caller to use specific trust roots for specific
rendezvous.
- Tests still to be modified.