QredoCrypto: push raw ed25519 down.
This does not exist at all in QC right now, and is implemented at the
CryptoImpl layer, which is not the correct layer of abstraction. All raw
algorithms belong in QredoCrypto.

This patch does not break the CryptoImpl Ed25519 interface yet, but “thins” it
out so that it does as little work as possible, before calling into QC.